Update types and severities according to Drupal numenclature
In Drupal, `type` is a value such as `access denied`, `cron` or `user`; 
`severity` is ` a value such as `Emergency`, `Alert`, `Debug`, etc, which are mapped to an ENUM, and can be translated to a value representing the severity. You can see this information in the Recent Log Messages pages in the reports menu.

In this plugin:
* `type` is the string representing the Drupal `severity`
* `severity` is the value representing the Drupal `severity`
* `Drupal severity` is nowhere to be found :(

With this commit, maybe it's a bit of breaking changes, but this way the nomenclature matches to the one in Drupal, plus we expose the `Drupal severity`, which is quite useful, for example in 404 cases, as we are only getting something like `Notice` but we don't know what exactly, just a URL in the message.

Because maybe someone would still need the `severity` in integer value, I added it in another attribute called `severity_code`.

Hope you can find it useful!